code tidying up
mainly with a view to pulling the bits that are now in the `util.atom`
package out into a separate library
