+++ app/model/ThriftUtil.scala
@@ -2,78 +2,13 @@ package model
 
 import com.gu.contentatom.thrift._
 import atom.media._
+
 import java.util.UUID.randomUUID
 import play.api.mvc.{ BodyParser, BodyParsers }
 import scala.concurrent.ExecutionContext
 
-trait AtomDataTyper[D] {
-  def getData(a: Atom): D
-  def setData(a: Atom, newData: D): Atom
-  def makeDefaultHtml(a: Atom): String
-}
-
-trait AtomImplicits[D] {
-  val dataTyper: AtomDataTyper[D]
-  implicit class AtomWithData(atom: Atom) {
-    def tdata = dataTyper.getData(atom)
-    def withData(data: D): Atom =
-      dataTyper.setData(atom, data).updateDefaultHtml
-    def updateData(f: D => D): Atom = withData(f(atom.tdata))
-    def withRevision(f: Long => Long): Atom = atom.copy(
-      contentChangeDetails = atom.contentChangeDetails.copy(
-        revision = f(atom.contentChangeDetails.revision)
-      )
-    )
-    def withRevision(newRevision: Long): Atom = withRevision(_ => newRevision)
-    def updateDefaultHtml = atom.copy(defaultHtml = dataTyper.makeDefaultHtml(atom))
-  }
-}
-
-trait MediaAtomImplicits extends AtomImplicits[MediaAtom] {
-  val dataTyper = new AtomDataTyper[MediaAtom] {
-    def getData(a: Atom) = a.data.asInstanceOf[AtomData.Media].media
-    def setData(a: Atom, newData: MediaAtom) =
-      a.copy(data = a.data.asInstanceOf[AtomData.Media].copy(media = newData))
-    def makeDefaultHtml(a: Atom) = {
-      val data = getData(a)
-      data.assets
-        .find(_.version == data.activeVersion)
-        .map(asset => views.html.MediaAtom.embedAsset(asset).toString)
-        .getOrElse(s"<div></div>")
-    }
-  }
-}
-
-object MediaAtomImplicits extends MediaAtomImplicits
-
 object ThriftUtil {
   type ThriftResult[A] = Either[String, A]
-
-  // implicit val mediaAtomDefaultHtml = new DefaultHTMLGenerator[AtomData.Media] {
-  //   def makeDefaultHtml(data: AtomData.Media) = {
-  //     data.media.assets
-  //       .map(asset => views.html.MediaAtom.embedAsset(asset))
-  //       .mkString("\n")
-  //   }
-  // }
-
-  // implicit class AtomDataWithType(a: Atom) {
-  //   def dataAs[D <: AtomData : Manifest]: D = a.data.asInstanceOf[D]
-
-  //   def mediaData = dataAs[AtomData.Media].media
-
-  //   def updateDataAs[D <: AtomData : Manifest](f: D => D): Atom = a.copy(data = f(a.dataAs[D]))
-
-  //   def updateMediaData(f: MediaAtom => MediaAtom) =
-  //     updateDataAs[AtomData.Media](atomData => atomData.copy(media = f(atomData.media)))
-
-  //   def withRevision(f: Long => Long): Atom = a.copy(
-  //     contentChangeDetails = a.contentChangeDetails.copy(
-  //       revision = f(a.contentChangeDetails.revision)
-  //     )
-  //   )
-  //   def withRevision(newRevision: Long): Atom = withRevision(_ => newRevision)
-  // }
 
   val youtube = "https?://www.youtube.com/watch\\?v=([^&]+)".r
