+++ app/views/MediaAtom/atomListTable.scala.html
@@ -1,5 +1,5 @@
 @import com.gu.contentatom.thrift._
-@import model.MediaAtomImplicits._
+@import util.atom.MediaAtomImplicits._
 @(atoms: TraversableOnce[Atom])
         <table id="atoms">
             <tr>
