+++ app/views/MediaAtom/atomTable.scala.html
@@ -1,5 +1,5 @@
 @import com.gu.contentatom.thrift._
-@import model.MediaAtomImplicits._
+@import util.atom.MediaAtomImplicits._
 @(atom: Atom)
         <h1>Atom Data</h1>
         <table id="atom-data">
