+++ app/views/MediaAtom/displayAtom.scala.html
@@ -1,6 +1,5 @@
 @import com.gu.contentatom.thrift._
-@import model.MediaAtomImplicits._
-@import util.MediaAtomUtil._
+@import util.atom.MediaAtomImplicits._
 @(atom: Atom)
 <html>
     <head>
